requirements.txt
Django==1.11.6
django-cors-headers==2.5.2
djangorestframework==3.9.2
gunicorn==19.7.1
pytz==2019.3
sqlparse==0.3.0
